Conflicts:
	src/common/kirin-gwt/pom.xml
	src/ios/KirinKit/KirinKit.xcodeproj/project.pbxproj
	src/ios/KirinKit/KirinKit/Core/KirinExtensions.m
	src/ios/KirinKit/KirinKit/Extensions/FileSystemBackend.m
	src/ios/KirinKit/KirinKit/Extensions/NewNotificationsImpl.m
	src/ios/KirinKit/KirinKit/Extensions/SettingsBackend.m

+++ src/ios/KirinKit/KirinKit/Extensions/GWTServices/NewNetworkingImpl.m
@@ -76,7 +76,7 @@
              }
              
              NSHTTPURLResponse * httpResp = (NSHTTPURLResponse*) response;
-             int respCode = httpResp.statusCode;
+             NSInteger respCode = httpResp.statusCode;
              NSDictionary* headerDict = [httpResp allHeaderFields];
              NSMutableArray *keyArr = [[NSMutableArray alloc] init];
              NSEnumerator *keyEnum = [headerDict keyEnumerator];
